requirements.txt
Django==1.8.5
django-annoying==0.8.4
django-bower==5.0.4
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-extensions==1.5.7
django-extra-views==0.7.1
django-grappelli==2.7.1
django-markdown-deux==1.0.5
django-pipeline==1.5.4
django-registration-redux==1.2
futures==3.0.3
jsonfield==1.0.3
markdown2==2.3.0
psycopg2==2.6.1
six==1.9.0
sqlparse==0.1.16
Werkzeug==0.10.4
wheel==0.24.0
